Group test_starccm and openfoam, and improve logging message output
- Move duplicated codes to common file
- Modify two tests' config files to make sure they are the same and can be used by both tests
- Add dependencies.install.sh in openfoam testdir as well to avoid failures
- Create a new shared_performance_test_cluster fixture, to group two tests in the same stack
- Now, if openfoam and starccm are not installed, it will not log errors
- Now, when percentage_difference is 0, will log matching baseline message
